+++ scanpipe/filters.py
@@ -40,10 +40,12 @@ from scanpipe.models import Run
 
 scanpipe_app = apps.get_app_config("scanpipe")
 
+PAGE_VAR = "page"
+
 
 class FilterSetUtilsMixin:
-    empty_value = "EMPTY"
-    other_value = "Other"
+    empty_value = "_EMPTY_"
+    other_value = "_OTHER_"
 
     @staticmethod
     def remove_field_from_query_dict(query_dict, field_name, remove_value=None):
@@ -97,6 +99,22 @@ class FilterSetUtilsMixin:
     def verbose_name_plural(cls):
         return cls.Meta.model._meta.verbose_name_plural
 
+    @property
+    def params(self):
+        return dict(self.data.items())
+
+    @property
+    def params_for_search(self):
+        """
+        Returns the current request query parameter used to keep the state
+        of the filters when using the search form.
+        The pagination and the search value is removed from those parameters.
+        """
+        params = self.params
+        params.pop(PAGE_VAR, None)
+        params.pop("search", None)
+        return params
+
     def filter_queryset(self, queryset):
         """
         Adds the ability to filter by empty and none values providing the "magic"
@@ -107,7 +125,7 @@ class FilterSetUtilsMixin:
             field_name = self.filters[name].field_name
             if value == self.empty_value:
                 queryset = queryset.filter(**{f"{field_name}__in": EMPTY_VALUES})
-            elif value == self.other_value:
+            elif value == self.other_value and hasattr(queryset, "less_common"):
                 return queryset.less_common(name)
             else:
                 queryset = self.filters[name].filter(queryset, value)