+++ scanpipe/models.py
@@ -76,8 +76,6 @@ from rq.job import JobStatus
 from scancodeio import __version__ as scancodeio_version
 from scanpipe import spdx
 from scanpipe import tasks
-from scanpipe.packagedb_models import AbstractPackage
-from scanpipe.packagedb_models import AbstractResource
 
 logger = logging.getLogger(__name__)
 scanpipe_app = apps.get_app_config("scanpipe")
@@ -109,6 +107,48 @@ class UUIDPKModel(models.Model):
     @property
     def short_uuid(self):
         return str(self.uuid)[0:8]
+
+
+class HashFieldsMixin(models.Model):
+    """
+    The hash fields are not indexed by default, use the `indexes` in Meta as needed:
+
+    class Meta:
+        indexes = [
+            models.Index(fields=['md5']),
+            models.Index(fields=['sha1']),
+            models.Index(fields=['sha256']),
+            models.Index(fields=['sha512']),
+        ]
+    """
+
+    md5 = models.CharField(
+        _("MD5"),
+        max_length=32,
+        blank=True,
+        help_text=_("MD5 checksum hex-encoded, as in md5sum."),
+    )
+    sha1 = models.CharField(
+        _("SHA1"),
+        max_length=40,
+        blank=True,
+        help_text=_("SHA1 checksum hex-encoded, as in sha1sum."),
+    )
+    sha256 = models.CharField(
+        _("SHA256"),
+        max_length=64,
+        blank=True,
+        help_text=_("SHA256 checksum hex-encoded, as in sha256sum."),
+    )
+    sha512 = models.CharField(
+        _("SHA512"),
+        max_length=128,
+        blank=True,
+        help_text=_("SHA512 checksum hex-encoded, as in sha512sum."),
+    )
+
+    class Meta:
+        abstract = True
 
 
 class AbstractTaskFieldsModel(models.Model):
@@ -1546,13 +1586,23 @@ class CodebaseResource(
     ScanFieldsModelMixin,
     ExtraDataFieldMixin,
     SaveProjectErrorMixin,
-    AbstractResource,
+    HashFieldsMixin,
+    models.Model,
 ):
     """
     A project Codebase Resources are records of its code files and directories.
     Each record is identified by its path under the project workspace.
+
+    These model fields should be kept in line with `scancode.resource.Resource`.
     """
 
+    path = models.CharField(
+        max_length=2000,
+        help_text=_(
+            "The full path value of a resource (file or directory) in the "
+            "archive it is from."
+        ),
+    )
     rootfs_path = models.CharField(
         max_length=2000,
         blank=True,
@@ -1566,6 +1616,11 @@ class CodebaseResource(
         blank=True,
         max_length=30,
         help_text=_("Analysis status for this resource."),
+    )
+    size = models.BigIntegerField(
+        blank=True,
+        null=True,
+        help_text=_("Size in bytes."),
     )
     tag = models.CharField(
         blank=True,
@@ -1644,7 +1699,6 @@ class CodebaseResource(
             "provided policies."
         ),
     )
-
     package_data = models.JSONField(
         default=list,
         blank=True,
@@ -1973,11 +2027,173 @@ class DiscoveredPackageQuerySet(PackageURLQuerySetMixin, ProjectRelatedQuerySet)
     pass
 
 
+class AbstractPackage(models.Model):
+    """
+    These model fields should be kept in line with `packagedcode.models.PackageData`.
+    """
+
+    filename = models.CharField(
+        max_length=255,
+        blank=True,
+        help_text=_(
+            "File name of a Resource sometimes part of the URI proper"
+            "and sometimes only available through an HTTP header."
+        ),
+    )
+    primary_language = models.CharField(
+        max_length=50,
+        blank=True,
+        help_text=_("Primary programming language."),
+    )
+    description = models.TextField(
+        blank=True,
+        help_text=_(
+            "Description for this package. "
+            "By convention the first line should be a summary when available."
+        ),
+    )
+    release_date = models.DateField(
+        blank=True,
+        null=True,
+        help_text=_(
+            "The date that the package file was created, or when "
+            "it was posted to its original download source."
+        ),
+    )
+    homepage_url = models.CharField(
+        _("Homepage URL"),
+        max_length=1024,
+        blank=True,
+        help_text=_("URL to the homepage for this package."),
+    )
+    download_url = models.CharField(
+        _("Download URL"),
+        max_length=2048,
+        blank=True,
+        help_text=_("A direct download URL."),
+    )
+    size = models.BigIntegerField(
+        blank=True,
+        null=True,
+        help_text=_("Size in bytes."),
+    )
+    bug_tracking_url = models.CharField(
+        _("Bug tracking URL"),
+        max_length=1024,
+        blank=True,
+        help_text=_("URL to the issue or bug tracker for this package."),
+    )
+    code_view_url = models.CharField(
+        _("Code view URL"),
+        max_length=1024,
+        blank=True,
+        help_text=_("a URL where the code can be browsed online."),
+    )
+    vcs_url = models.CharField(
+        _("VCS URL"),
+        max_length=1024,
+        blank=True,
+        help_text=_(
+            "A URL to the VCS repository in the SPDX form of: "
+            '"git", "svn", "hg", "bzr", "cvs", '
+            "https://github.com/nexb/scancode-toolkit.git@405aaa4b3 "
+            'See SPDX specification "Package Download Location" '
+            "at https://spdx.org/spdx-specification-21-web-version#h.49x2ik5"
+        ),
+    )
+    repository_homepage_url = models.CharField(
+        _("Repository homepage URL"),
+        max_length=1024,
+        blank=True,
+        help_text=_(
+            "URL to the page for this package in its package repository. "
+            "This is typically different from the package homepage URL proper."
+        ),
+    )
+    repository_download_url = models.CharField(
+        _("Repository download URL"),
+        max_length=1024,
+        blank=True,
+        help_text=_(
+            "Download URL to download the actual archive of code of this "
+            "package in its package repository. "
+            "This may be different from the actual download URL."
+        ),
+    )
+    api_data_url = models.CharField(
+        _("API data URL"),
+        max_length=1024,
+        blank=True,
+        help_text=_(
+            "API URL to obtain structured data for this package such as the "
+            "URL to a JSON or XML api its package repository."
+        ),
+    )
+    copyright = models.TextField(
+        blank=True,
+        help_text=_("Copyright statements for this package. Typically one per line."),
+    )
+    license_expression = models.TextField(
+        blank=True,
+        help_text=_(
+            "The normalized license expression for this package as derived "
+            "from its declared license."
+        ),
+    )
+    declared_license = models.TextField(
+        blank=True,
+        help_text=_(
+            "The declared license mention or tag or text as found in a "
+            "package manifest."
+        ),
+    )
+    notice_text = models.TextField(
+        blank=True,
+        help_text=_("A notice text for this package."),
+    )
+    manifest_path = models.CharField(
+        max_length=1024,
+        blank=True,
+        help_text=_(
+            "A relative path to the manifest file if any, such as a "
+            "Maven .pom or a npm package.json."
+        ),
+    )
+    contains_source_code = models.BooleanField(null=True, blank=True)
+    datasource_id = models.CharField(
+        max_length=64,
+        blank=True,
+        help_text=_(
+            "The identifier for the datafile handler used to obtain this package."
+        ),
+    )
+    file_references = models.JSONField(
+        default=list,
+        blank=True,
+        help_text=_(
+            "List of file paths and details for files referenced in a package "
+            "manifest. These may not actually exist on the filesystem. "
+            "The exact semantics and base of these paths is specific to a "
+            "package type or datafile format."
+        ),
+    )
+    parties = models.JSONField(
+        default=list,
+        blank=True,
+        help_text=_("A list of parties such as a person, project or organization."),
+    )
+
+    class Meta:
+        abstract = True
+
+
 class DiscoveredPackage(
     ProjectRelatedModel,
     ExtraDataFieldMixin,
     SaveProjectErrorMixin,
     UpdateFromDataMixin,
+    HashFieldsMixin,
+    PackageURLMixin,
     AbstractPackage,
 ):
     """
@@ -1989,6 +2205,9 @@ class DiscoveredPackage(
     See https://github.com/package-url for more details.
     """
 
+    uuid = models.UUIDField(
+        verbose_name=_("UUID"), default=uuid.uuid4, unique=True, editable=False
+    )
     codebase_resources = models.ManyToManyField(
         "CodebaseResource", related_name="discovered_packages"
     )
@@ -2000,8 +2219,6 @@ class DiscoveredPackage(
         db_index=True,
         help_text=_("Unique identifier for this package."),
     )
-
-    # `AbstractPackage` model overrides:
     keywords = models.JSONField(default=list, blank=True)
     source_packages = models.JSONField(default=list, blank=True)
 
@@ -2009,6 +2226,15 @@ class DiscoveredPackage(
 
     class Meta:
         ordering = ["uuid"]
+        indexes = [
+            models.Index(fields=["filename"]),
+            models.Index(fields=["primary_language"]),
+            models.Index(fields=["size"]),
+            models.Index(fields=["md5"]),
+            models.Index(fields=["sha1"]),
+            models.Index(fields=["sha256"]),
+            models.Index(fields=["sha512"]),
+        ]
         constraints = [
             models.UniqueConstraint(
                 fields=["project", "package_uid"],